CPM-64;Feat;Moves TreeView component out of TreePanel
