+++ src/components/tree-view/index.ts
@@ -1,0 +1,2 @@
+import TreeView from './tree-view'
+export default TreeView
